+++ src/StackExchange.Redis/Enums/RedisCommand.cs
@@ -60,7 +60,6 @@ internal enum RedisCommand
     GEOSEARCH,
     GEOSEARCHSTORE,
 
-    GCRA,
     GET,
     GETBIT,
     GETDEL,
@@ -326,7 +325,6 @@ internal static class RedisCommandExtensions
             case RedisCommand.EXPIREAT:
             case RedisCommand.FLUSHALL:
             case RedisCommand.FLUSHDB:
-            case RedisCommand.GCRA:
             case RedisCommand.GEOSEARCHSTORE:
             case RedisCommand.GETDEL:
             case RedisCommand.GETEX: