+++ src/StackExchange.Redis/Enums/SortedSetWhen.cs
@@ -36,15 +36,6 @@ namespace StackExchange.Redis
 
     internal static class SortedSetWhenExtensions
     {
-        internal static uint CountBits(this SortedSetWhen when)
-        {
-            uint v = (uint)when;
-            v -= (v >> 1) & 0x55555555; // reuse input as temporary
-            v = (v & 0x33333333) + ((v >> 2) & 0x33333333); // temp
-            uint c = ((v + (v >> 4) & 0xF0F0F0F) * 0x1010101) >> 24; // count
-            return c;
-        }
-
         internal static SortedSetWhen Parse(When when) => when switch
         {
             When.Always => SortedSetWhen.Always,