+++ tests/StackExchange.Redis.Tests/GarbageCollectionTests.cs
@@ -38,12 +38,17 @@ public class GarbageCollectionTests(ITestOutputHelper helper) : TestBase(helper)
         var wr = new WeakReference(conn);
         conn = null;
 
-        ForceGC();
-        await Task.Delay(2000).ForAwait(); // GC is twitchy
-        ForceGC();
+        for (int i = 0; i < 5 && wr.IsAlive; i++)
+        {
+            ForceGC();
+            await Task.Delay(2000).ForAwait(); // GC is twitchy
+            ForceGC();
+        }
 
         // should be collectable
         Assert.Null(wr.Target);
+        // just to ensure we wrote conn, and to suppress a warning
+        Assert.Null(conn);
 
 // #if DEBUG // this counter only exists in debug
 //            int after = ConnectionMultiplexer.CollectedWithoutDispose;